Fix `pullCommand` backspace logic
When the command-line mode was being exited with `<bs>`,
the old code was consuming the printable key event in the same iteration
it was setting the `finish` flag.
Now printable characters generate only one entry, so a fake event is
prepended.

TODO actually move code around or add additional condition levels
to avoid unnecessary typeahead modifications.